+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# A library that provides functionality to the @RandomUtils_bot
+# A library that provides functionality to the @randomutils_bot
 # Copyright (C) 2017-2018
 # Javier Gines Sanchez <software@javisite.com>
 #
@@ -15,11 +15,20 @@ All notable changes will be documented in this file.
 - Command /anonymousMessage (Speak?¿)
 - Command /note
 
+## [0.0.4 - Snapshot.1a] - 2018-02-21
+### Added
+- Added inline shortlink command
+
+### Changed
+- Better detection of links in shortlink
+
+
 ## [0.0.3 - Snapshot.1e] - 2018-02-18
 ### Added
 - Command /shortLink
 - Added /help text
 - Add /random command to Inline
+- Short Commands
 
 ### Changed
 - Tokens reading